Add custom domain redirection support
